pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ClickFlow-Sync"
version = "1.1.0"
description = "Professional ClickUp task automation with state sync."
readme = "README.md"
authors = [{name = "Jenil Mistry"}]
dependencies = [
    "requests",
    "python-dotenv"
]

[tool.setuptools.packages.find]
where = ["src"]